Removing 'throw exception=true' nlog adjustment for tests, as it was causing a test to fail that we did not think was worth spending time fixing
